refactor: move Location to separate file
